Extract navigation bar to a separate component for Hub App layouts
